view-time (5)/backend/requirements.txt
# Workspace version: 1.47.0
databutton==0.38.34
ipykernel==6.29.5
fastapi==0.115.12
python-multipart==0.0.9
uvicorn[standard]==0.34.2

openai
beautifulsoup4
requests
google-auth
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-cloud-firestore
